Harmonise displaying station names with the help of forDisplay computed attribute
